# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JaIre Web3 Coworking Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BaireChatRequest {
  message: string;
  session_id?: string;
  user_id?: string;
}

export type WorkspaceWorkspaceType =
  (typeof WorkspaceWorkspaceType)[keyof typeof WorkspaceWorkspaceType];

export const WorkspaceWorkspaceType = {
  hot_desk: "hot_desk",
  private_suite: "private_suite",
  meeting_room: "meeting_room",
  lounge: "lounge",
} as const;

export interface Workspace {
  id: string;
  name: string;
  description?: string;
  capacity?: number;
  /** Hourly rate in Nigerian Naira */
  hourly_rate_ngn: number;
  /** Hourly rate in USDC */
  hourly_rate_usdc: number;
  amenities?: string[];
  image_url?: string;
  is_available: boolean;
  current_occupant_count?: number;
  floor?: number;
  workspace_type?: WorkspaceWorkspaceType;
}

export interface BaireChatResponse {
  reply: string;
  session_id: string;
  suggested_actions?: string[];
  workspace_recommendations?: Workspace[];
}

export interface BaireVoiceResponse {
  transcript: string;
  reply: string;
  audio_url?: string;
  session_id: string;
}

export interface TimeSlot {
  start_time: string;
  end_time: string;
  is_available: boolean;
  booking_id?: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type BookingPaymentMethod =
  (typeof BookingPaymentMethod)[keyof typeof BookingPaymentMethod];

export const BookingPaymentMethod = {
  paystack: "paystack",
  roqqu: "roqqu",
  usdc_wallet: "usdc_wallet",
} as const;

export interface Booking {
  id: string;
  workspace_id: string;
  workspace_name?: string;
  user_id?: string;
  status: BookingStatus;
  check_in_time?: string;
  check_out_time?: string;
  planned_duration_hours?: number;
  actual_duration_seconds?: number;
  escrow_amount_usdc?: number;
  billed_amount_usdc?: number;
  refunded_amount_usdc?: number;
  /** Solana escrow account pubkey */
  escrow_account?: string;
  transaction_signature?: string;
  payment_method?: BookingPaymentMethod;
  ngn_amount_paid?: number;
}

export type CreateBookingBodyPaymentMethod =
  (typeof CreateBookingBodyPaymentMethod)[keyof typeof CreateBookingBodyPaymentMethod];

export const CreateBookingBodyPaymentMethod = {
  paystack: "paystack",
  roqqu: "roqqu",
  usdc_wallet: "usdc_wallet",
} as const;

export interface CreateBookingBody {
  workspace_id: string;
  planned_duration_hours: number;
  payment_method?: CreateBookingBodyPaymentMethod;
  web3auth_token?: string;
  user_email?: string;
  user_name?: string;
}

export interface CheckoutSummary {
  booking_id: string;
  workspace_name?: string;
  duration_seconds: number;
  duration_display?: string;
  billed_usdc: number;
  refunded_usdc?: number;
  yield_earned_usdc?: number;
  transaction_signature?: string;
  check_in_time?: string;
  check_out_time?: string;
}

export type SessionStatusStatus =
  (typeof SessionStatusStatus)[keyof typeof SessionStatusStatus];

export const SessionStatusStatus = {
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface SessionStatus {
  booking_id: string;
  workspace_id?: string;
  workspace_name?: string;
  status: SessionStatusStatus;
  elapsed_seconds: number;
  planned_seconds?: number;
  elapsed_display?: string;
  current_cost_usdc: number;
  current_cost_ngn?: number;
  escrowed_usdc?: number;
  check_in_time?: string;
}

export interface WalletConnectBody {
  web3auth_token: string;
  email?: string;
}

export type WalletInfoWalletType =
  (typeof WalletInfoWalletType)[keyof typeof WalletInfoWalletType];

export const WalletInfoWalletType = {
  mpc: "mpc",
  custodial: "custodial",
  external: "external",
} as const;

export interface WalletInfo {
  public_key: string;
  wallet_type?: WalletInfoWalletType;
  balance_usdc?: number;
  balance_sol?: number;
  verifier_id?: string;
  connected: boolean;
}

export interface WalletBalance {
  public_key: string;
  balance_usdc: number;
  balance_sol: number;
  last_updated?: string;
}

export type FundWalletBodyPaymentMethod =
  (typeof FundWalletBodyPaymentMethod)[keyof typeof FundWalletBodyPaymentMethod];

export const FundWalletBodyPaymentMethod = {
  paystack: "paystack",
  roqqu: "roqqu",
} as const;

export interface FundWalletBody {
  /** Amount in Nigerian Naira to convert to USDC */
  amount_ngn: number;
  payment_method: FundWalletBodyPaymentMethod;
  email: string;
}

export interface FundWalletResponse {
  /** Paystack/Roqqu checkout URL */
  payment_url: string;
  reference: string;
  amount_ngn: number;
  estimated_usdc: number;
  /** NGN per USDC */
  exchange_rate?: number;
}

export type AnalyticsDashboardRevenueTrend =
  (typeof AnalyticsDashboardRevenueTrend)[keyof typeof AnalyticsDashboardRevenueTrend];

export const AnalyticsDashboardRevenueTrend = {
  up: "up",
  down: "down",
  flat: "flat",
} as const;

export interface WorkspaceStat {
  workspace_id: string;
  workspace_name: string;
  sessions_today: number;
  revenue_today_usdc?: number;
  occupancy_rate?: number;
}

export interface AnalyticsDashboard {
  total_revenue_usdc: number;
  total_revenue_ngn?: number;
  total_sessions_today?: number;
  active_sessions: number;
  /** 0-1 fraction */
  occupancy_rate: number;
  avg_session_duration_hours?: number;
  top_workspace?: string;
  /** Kamino yield earned */
  yield_earned_usdc?: number;
  revenue_trend?: AnalyticsDashboardRevenueTrend;
  revenue_change_pct?: number;
  workspace_stats?: WorkspaceStat[];
}

export interface OccupancyData {
  date: string;
  workspace_id: string;
  workspace_name?: string;
  occupancy_rate: number;
  total_hours?: number;
}

export interface RevenueData {
  date: string;
  workspace_id: string;
  workspace_name?: string;
  revenue_usdc: number;
  revenue_ngn?: number;
  session_count?: number;
}

export type ActivityEventEventType =
  (typeof ActivityEventEventType)[keyof typeof ActivityEventEventType];

export const ActivityEventEventType = {
  check_in: "check_in",
  check_out: "check_out",
  payment: "payment",
  wallet_funded: "wallet_funded",
  booking_cancelled: "booking_cancelled",
} as const;

export type ActivityEventMetadata = { [key: string]: unknown };

export interface ActivityEvent {
  id: string;
  event_type: ActivityEventEventType;
  workspace_id?: string;
  workspace_name?: string;
  user_name?: string;
  amount_usdc?: number;
  amount_ngn?: number;
  timestamp: string;
  metadata?: ActivityEventMetadata;
}

export type BaireVoiceBody = {
  audio: Blob;
  session_id?: string;
};

export type ListWorkspacesParams = {
  /**
   * Filter by availability date (YYYY-MM-DD)
   */
  date?: string;
  /**
   * Minimum hourly rate (NGN)
   */
  min_rate?: number;
  /**
   * Maximum hourly rate (NGN)
   */
  max_rate?: number;
  /**
   * Comma-separated amenities filter
   */
  amenities?: string;
};

export type GetWorkspaceAvailabilityParams = {
  date: string;
};

export type ListBookingsParams = {
  status?: ListBookingsStatus;
  limit?: number;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
  pending: "pending",
} as const;

export type GetOccupancyAnalyticsParams = {
  days?: number;
};

export type GetRevenueAnalyticsParams = {
  days?: number;
};

export type GetActivityFeedParams = {
  limit?: number;
};