# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JaIre Web3 Coworking Platform API
servers:
  - url: /api
    description: Baire AI concierge (Node api-server)
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List all available workspaces
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: Filter by availability date (YYYY-MM-DD)
        - name: min_rate
          in: query
          schema:
            type: number
          description: Minimum hourly rate (NGN)
        - name: max_rate
          in: query
          schema:
            type: number
          description: Maximum hourly rate (NGN)
        - name: amenities
          in: query
          schema:
            type: string
          description: Comma-separated amenities filter
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace details
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workspace detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          description: Not found
  /workspaces/{workspaceId}/availability:
    get:
      operationId: getWorkspaceAvailability
      tags: [workspaces]
      summary: Get time slots for a workspace on a given date
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Availability slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"
...
components:
  schemas:
...
    Workspace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        capacity:
          type: integer
        hourly_rate_ngn:
          type: number
          description: Hourly rate in Nigerian Naira
        hourly_rate_usdc:
          type: number
          description: Hourly rate in USDC
        amenities:
          type: array
          items:
            type: string
        image_url:
          type: string
        is_available:
          type: boolean
        current_occupant_count:
          type: integer
        floor:
          type: integer
        workspace_type:
          type: string
          enum: [hot_desk, private_suite, meeting_room, lounge]
      required: [id, name, hourly_rate_ngn, hourly_rate_usdc, is_available]
    TimeSlot:
      type: object
      properties:
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        is_available:
          type: boolean
        booking_id:
          type: string
      required: [start_time, end_time, is_available]
...