# lib/api-zod/src/generated/types/workspace.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JaIre Web3 Coworking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkspaceWorkspaceType } from "./workspaceWorkspaceType";

export interface Workspace {
  id: string;
  name: string;
  description?: string;
  capacity?: number;
  /** Hourly rate in Nigerian Naira */
  hourly_rate_ngn: number;
  /** Hourly rate in USDC */
  hourly_rate_usdc: number;
  amenities?: string[];
  image_url?: string;
  is_available: boolean;
  current_occupant_count?: number;
  floor?: number;
  workspace_type?: WorkspaceWorkspaceType;
}